// src/app/classes/books.interface.ts
export interface BooksGeneral {
    nombre: string;
    descripcion: string;
    img: string;
    publicacion: string;
    editorial: string;
    autor: string;
    precio: string;
    categoria: string;
};
export interface BooksNovelas {
    nombre: string;
    descripcion: string;
    img: string;
    publicacion: string;
    editorial: string;
    autor: string;
    precio: string;
    categoria: string;
};
export interface BooksFiccion {
    nombre: string;
    descripcion: string;
    img: string;
    publicacion: string;
    editorial: string;
    autor: string;
    precio: string;
    categoria: string;
};